requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asgiref=3.7.2=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
django=5.0=pypi_0
django-cors-headers=4.3.1=pypi_0
django-filter=23.5=pypi_0
djangorestframework=3.14.0=pypi_0
expat=2.5.0=h6a678d5_0
idna=3.6=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.5.1=pypi_0
ncurses=6.4=h6a678d5_0
openssl=3.0.12=h7f8727e_0
pip=23.3.1=py312h06a4308_0
python=3.12.0=h996f2a0_0
pytz=2023.3.post1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
setuptools=68.0.0=py312h06a4308_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
tk=8.6.12=h1ccaba5_0
tzdata=2023c=h04d1e81_0
urllib3=2.1.0=pypi_0
wheel=0.41.2=py312h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
